moves files to the src folder
