+++ src/main.js
@@ -1,6 +1,6 @@
 // Main entry point
 import { initBurger } from './burger/burger.js';
-import './partials/filter_panels/filter_panels.js';
+import './filter_panels/filter_panels.js';
 function start() {
   initBurger();
 }
